simulator/requirements.txt
# Simulator runtime dependencies
pika>=1.3.2,<2.0
pyyaml>=6.0.1,<7.0
pydantic>=2.5.0,<3.0
prometheus-client>=0.20.0,<1.0

# Testing
pytest>=8.0.0,<9.0
pytest-cov>=4.1.0,<6.0
